2011: split up KeyringSetPassword.tsx and KeyringChangePassword.tsx into two component files
